Remove FullyQualifiedClassNameConverter
It's better implemented as a method on FullyQualifiedClassName.
Closes #84.
